% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{colorizeLAS}
\alias{colorizeLAS}
\title{Colorize LAS point cloud by attribute using custom palette}
\usage{
colorizeLAS(las, attribute, palette_func, n = 100)
}
\arguments{
\item{las}{LAS object to colorize}

\item{attribute}{character. Name of the attribute column to color by}

\item{palette_func}{function. Color palette function (e.g., rainbow, heat.colors, terrain.colors)}

\item{n}{integer. Number of colors in the palette. Default is 100.}
}
\value{
LAS object with R, G, B columns added/updated
}
\description{
This function assigns RGB colors to a LAS point cloud based on a specified attribute
and color palette function. It normalizes the attribute values and maps them to colors.
}
\examples{
\dontrun{
# Colorize by Z values using rainbow palette
las_colored <- colorizeLAS(las, "Z", rainbow)

# Colorize by intensity using heat colors
las_colored <- colorizeLAS(las, "Intensity", heat.colors, n = 50)

# Use with RColorBrewer
library(RColorBrewer)
las_colored <- colorizeLAS(las, "Classification", 
                          function(n) brewer.pal(min(n, 9), "Set1"))
}
}
