% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,LAS,missing-method}
\alias{plot,LAS,missing-method}
\title{Display big 3D point clouds}
\usage{
\S4method{plot}{LAS,missing}(x, y, ...)
}
\arguments{
\item{x}{a point cloud with minimally 3 columns named X,Y,Z}

\item{y}{Unused (inherited from R base)}
}
\description{
Display arbitrary large in memory 3D point clouds from the lidR package. Keyboard can be use
to control the rendering
\itemize{
\item Rotate with left mouse button
\item Zoom with mouse wheel
\item Pan with right mouse button
\item Keyboard \if{html}{\out{<kbd>}}r\if{html}{\out{</kbd>}} or \if{html}{\out{<kbd>}}g\if{html}{\out{</kbd>}} or \if{html}{\out{<kbd>}}b\if{html}{\out{</kbd>}} to color with RGB
\item Keyboard \if{html}{\out{<kbd>}}z\if{html}{\out{</kbd>}} to color with Z
\item Keyboard \if{html}{\out{<kbd>}}i\if{html}{\out{</kbd>}} to color with Intensity
\item Keyboard \if{html}{\out{<kbd>}}c\if{html}{\out{</kbd>}} to color with Classification
\item Keyboard \if{html}{\out{<kbd>}}+\if{html}{\out{</kbd>}} or \if{html}{\out{<kbd>}}-\if{html}{\out{</kbd>}} to change the point size
\item Keyboard \if{html}{\out{<kbd>}}l\if{html}{\out{</kbd>}} to enable/disable eyes-dome lightning
}
}
