% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot,LAS,missing-method}
\alias{plot_xyzrgb}
\title{Display big 3D point clouds}
\usage{
plot(x, y, ...)

\S4method{plot}{LAS,missing}(x, y, ...)

plot_xyzrgb(x, y, z, r, g, b, id = NULL, size = 4)
}
\arguments{
\item{x}{A \code{LAS*} object from the lidR package}

\item{y}{Unused (inherited from R base)}
}
\description{
Display aritrary large in memory 3D point clouds. This function does no lag and can display
almost instantly millions of points. It is optimized to used the lowest possible memory
and is capable of displaying gigabytes of points allocating only few megabyte of memory
}
