% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_xyz}
\alias{plot_xyz}
\title{Display big 3D point clouds}
\usage{
plot_xyz(x, y, z, col, size = 2)
}
\arguments{
\item{x}{vector of x coordinates}

\item{y}{vector of y coordinates}

\item{z}{vector of z coordinates}

\item{col}{matrix n x 3 of rgb values to color the points}

\item{size}{the size of the points}
}
\description{
Display big 3D point cloud using rgl style but is memory efficent and does no lag
even for millions of points.
}
\examples{
\dontrun{
x = runif(1000, 0, 100)
y = runif(1000, 0, 100)
z = runif(1000, 0, 100)
c = t(col2rgb(sample(rainbow(10), 1000,replace = T)))
plot_xyz(x,y,z, c)
}
}
