% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_xyzrgb}
\alias{plot_xyzrgb}
\title{Deprecated backward compatible function}
\usage{
plot_xyzrgb(x, y, z, r, g, b, id = NULL, size = 4)
}
\arguments{
\item{x, y, z}{numeric vector}

\item{r, g, b}{integer vector}

\item{id}{index}

\item{size}{not used}
}
\description{
Deprecated backward compatible function
}
